# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "My Project API",
        "title": "My Project API",
        "termsOfService": "https://github.com/inhere",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/inhere",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/books": {
            "post": {
                "description": "insert book booking data",
                "tags": [
                    "BooksApi"
                ],
                "summary": "Create a new book booking",
                "parameters": [
                    {
                        "description": "new booking data",
                        "name": "bodyData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.CreateBookingForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Need booking data!!",
                        "schema": {
                            "$ref": "#/definitions/model.JsonMapData"
                        }
                    },
                    "404": {
                        "description": "Cannot insert booking data",
                        "schema": {
                            "$ref": "#/definitions/model.JsonMapData"
                        }
                    }
                }
            }
        },
        "/books/{subject}": {
            "get": {
                "description": "get book details",
                "tags": [
                    "BooksApi"
                ],
                "summary": "Get multiple book details per page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book subject",
                        "name": "subject",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Need book subject",
                        "schema": {
                            "$ref": "#/definitions/model.JsonMapData"
                        }
                    },
                    "404": {
                        "description": "Can't find book subject",
                        "schema": {
                            "$ref": "#/definitions/model.JsonMapData"
                        }
                    }
                }
            }
        },
        "/config": {
            "get": {
                "tags": [
                    "InternalApi"
                ],
                "summary": "Get app config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "config key string",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "get app health",
                "tags": [
                    "InternalApi"
                ],
                "summary": "Health check",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "form.CreateBookingForm": {
            "type": "object",
            "required": [
                "user-id"
            ],
            "properties": {
                "pick-up-schedule": {
                    "description": "Pick up schedule",
                    "type": "string",
                    "example": "12-12-2012 12:12:12"
                },
                "title": {
                    "description": "BookID to be booked",
                    "type": "string",
                    "example": "harry potter"
                },
                "user-id": {
                    "description": "UserID who's booking",
                    "type": "integer"
                }
            }
        },
        "model.JsonMapData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}